Move the export_meeting into an own file.
